Move specialized qglue parsers to be alongside their matching data factories
