+++ glue/core/data_factories/fits.py
@@ -6,7 +6,7 @@ from collections import OrderedDict
 
 from glue.core.coordinates import coordinates_from_header, WCSCoordinates
 from glue.core.data import Component, Data
-from glue.config import data_factory
+from glue.config import data_factory, qglue_parser
 
 
 __all__ = ['is_fits', 'fits_reader', 'is_casalike', 'casalike_cube']
@@ -188,3 +188,15 @@ def casalike_cube(filename, **kwargs):
     for i in range(array.shape[0]):
         result.add_component(array[[i]], label='STOKES %i' % i)
     return result
+
+
+try:
+    from astropy.io.fits import HDUList
+except ImportError:
+    pass
+else:
+    # Put HDUList parser before list parser
+    @qglue_parser(HDUList, priority=100)
+    def _parse_data_hdulist(data, label):
+        from glue.core.data_factories.fits import fits_reader
+        return fits_reader(data, label=label)
