+++ glue/plugins/data_factories/spectral_cube/__init__.py
@@ -1,2 +1,2 @@
 def setup():
-    from .spectral_cube import spectral_cube_reader
+    from .spectral_cube import read_spectral_cube, parse_spectral_cube
