+++ glue/plugins/data_factories/spectral_cube/spectral_cube.py
@@ -3,9 +3,11 @@ from __future__ import absolute_import, division, print_function
 from spectral_cube import SpectralCube, StokesSpectralCube
 
 from glue.core import Data
-from glue.config import data_factory
+from glue.config import data_factory, qglue_parser
 from glue.core.data_factories.fits import is_fits
 from glue.core.coordinates import coordinates_from_wcs
+
+__all__ = ['read_spectral_cube', 'parse_spectral_cube']
 
 
 def is_spectral_cube(filename, **kwargs):
@@ -40,10 +42,15 @@ def spectral_cube_to_data(cube, label=None):
 
 
 @data_factory(label='FITS Spectral Cube', identifier=is_spectral_cube)
-def spectral_cube_reader(filename, **kwargs):
+def read_spectral_cube(filename, **kwargs):
     """
     Read in a FITS spectral cube. If multiple Stokes components are present,
     these are split into separate glue components.
     """
     cube = StokesSpectralCube.read(filename)
     return spectral_cube_to_data(cube)
+
+
+@qglue_parser((SpectralCube, StokesSpectralCube))
+def parse_spectral_cube(cube, label):
+    return [spectral_cube_to_data(cube, label=label)]
